requirements.txt
absl-py==1.3.0
Pillow==9.4.0
